separated add, back, next sub commands into individual classes to remove if else clutter
